Updating python setup.py line for singlegpu
The --singlegpu line when passed to build.sh is only properly
acknowledged and used in the libcuml build step. The cuml build
which is built with python does not use singlegpu. This commit
fixes that by adding -DSINGLEGPU=ON to the CUML_EXTRA_CMAKE_ARGS
before python setup.py runs.